# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Mailflow API
  version: 0.1.0
  description: Stable public contract shared by the web, desktop and future Swift clients.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /translations/{locale}:
    get:
      operationId: getTranslations
      security: []
      parameters:
        - name: locale
          in: path
          required: true
          schema: { type: string, enum: [en, es] }
      responses:
        "200": { description: Translation catalog with English fallback }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...